prevent access to another user's LPAs
